refactor(blog): organizar la feature blog en subcarpetas (feed, composer, comments, shared)
