+++ src/pages/[id]/blog.astro
@@ -10,7 +10,7 @@ import ContainerLink from '@/shared/components/ContainerLink.astro';
 import IconDocument from '@/shared/components/icons/IconDocument.astro';
 import IconUniversity from '@/shared/components/icons/IconUniversity.astro';
 import IconBlog from '@/shared/components/icons/IconBlog.astro';
-import BlogView from '@/features/blog/components/BlogView';
+import BlogView from '@/features/blog/feed/BlogView';
 import { getSubjectById } from '@/shared/services/api';
 
 const { id } = Astro.params;
